Remove yjit/src/external.c
